requirements.txt
ffmpeg==1.4
imageio==2.9.0
lpips==0.1.4
matplotlib==3.5.2
opencv-contrib-python==4.5.5.64
pyyaml==6.0
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.0
tensorboard==2.5.0
tqdm==4.61.2
pytorch==1.10.0
torchvision==0.11.1